Fixed use of article "the" in documentation
In the vim-r-plugin documentation the phrase "The Vim-R-plugin" would
occur often. This was replaced in the nvim-r documentation with "The
Nvim-R" but that is distracting. One would expect: "The Nvim-R plugin
this or that" or "Nvim-R this or that" (and not "the Nvim-R this or
that").

So Nvim-R is used as its name where are appropiate, and "the" remains
where there is a reference to something else (the Nvim-R session, or the
Nvim-R features, etc.).