Fixes #14 regular expression to identify triple back-ticks too greedy
Note that the lines above looking for double dots (e.g. line 1132)

    let lines = map(copy(lines), 'substitute(v:val, "^\\.\\. \\?", "", "")')

are left unchanged, simply because I do not know yet what the double
dots would represent in markdown. But a similar issue might occure there
is someone had a variable name starting with a dot (not common style,
but allowed in R).