Allows RViewDF to open in splits
- Adds a 'location' argument to the `nvim_viewdf` R function, which
specifies where the buffer showing the df will be opened e.g. 'tabnew',
'split', 'vsplit'. This argument is then passed to the `RViewDF` vimscript
function. A default value of 'tabnew' is used for consistency with the
current behaviour.

- Adds an 'nrows' argument to `nvim_viewdf`. If provided, `head(x, nrows)` is
written and opened as a csv, rather than the whole file.

- Modifies RAction so that a string with these new arguments in can be
passed to `nvim_viewdf`.

- Modifies RViewDF so that the data is opened in the given location.

- Adds mappings for opening in a split, vsplit and also opening the head
i.e. first 6 rows in a split above the current window. Also creates menu
items for these mappings in gvim.