RCreateMaps: Create plug mappings unconditionally
Previously, Plug mappings were only created if they were already in use
- i.e., if someone had mapped them to something in their .vimrc. The
problem with that is if someone wants to create buffer-local maps to an
nvim-R <Plug> mapping, they cannot do so from .vimrc. They must do so
once the buffer is loaded - ie within vim, after loading .vimrc.
This is usually done with an filetype autocmd.

See #470 for more