RCreateMaps: Create plug mappings unconditionally (fixes #470, fixes #494) (#472)
* RCreateMaps: Create plug mappings unconditionally

Previously, Plug mappings were only created if they were already in use
- i.e., if someone had mapped them to something in their .vimrc. The
problem with that is if someone wants to create buffer-local maps to an
nvim-R <Plug> mapping, they cannot do so from .vimrc. They must do so
once the buffer is loaded - ie within vim, after loading .vimrc.
This is usually done with an filetype autocmd.

See #470 for more

* Update mapping documentation to suggest buffer-local mappings

* Link to option to only use custom user settings

* Clarify purpose of RStop

The word 'stop' is a bit misleading, because it suggests killing the R
process totally. That doesn't always happen, it just interrupts.

* Fix filetype autocmd in documentation

* Flip check on user_maps_only

* Restore check to not map default if custom mapping exists

* Fix delay when RSendLine plug mapping is used

Co-authored-by: bluedrink9 <bluedrink9@github.com>