Partial fix for  #816
Creates a separate trigger for R_assign = 2 so that it doesn't drop out of insert mode unless two underscores (or R_assign_map characters) are typed. Consequently, inserts like "this_variable_name" will be correctly saved to the repeat register.

However the repeat register (@.) is still reset if a substitution is made. This will be harder to fix.